Minor update, upload model
