+++ backend/model/README
@@ -1,0 +1,5 @@
+Accurate universal English model (both for callcenter and wideband)
+
+Based on Appen Kaldi model https://github.com/Appen/UHV-OTS-Speech
+
+Dynamic graph version
