+++ backend/model/conf/mfcc.conf
@@ -1,0 +1,9 @@
+--sample-frequency=16000
+--use-energy=false
+--num-mel-bins=40
+--num-ceps=40
+--low-freq=20
+--high-freq=7600
+--allow-upsample=true
+--allow-downsample=true
+
