+++ backend/model/conf/model.conf
@@ -1,0 +1,10 @@
+--min-active=200
+--max-active=7000
+--beam=13.0
+--lattice-beam=6.0
+--acoustic-scale=1.0
+--frame-subsampling-factor=3
+--endpoint.silence-phones=1:2:3:4:5:11:12:13:14:15
+--endpoint.rule2.min-trailing-silence=0.5
+--endpoint.rule3.min-trailing-silence=1.0
+--endpoint.rule4.min-trailing-silence=2.0
