+++ backend/model/graph/phones.txt
@@ -1,0 +1,267 @@
+<eps> 0
+SIL 1
+SIL_B 2
+SIL_E 3
+SIL_I 4
+SIL_S 5
+LAUGHTER 6
+LAUGHTER_B 7
+LAUGHTER_E 8
+LAUGHTER_I 9
+LAUGHTER_S 10
+NOISE 11
+NOISE_B 12
+NOISE_E 13
+NOISE_I 14
+NOISE_S 15
+OOV 16
+OOV_B 17
+OOV_E 18
+OOV_I 19
+OOV_S 20
+SPN 21
+SPN_B 22
+SPN_E 23
+SPN_I 24
+SPN_S 25
+BRH 26
+BRH_B 27
+BRH_E 28
+BRH_I 29
+BRH_S 30
+CGH 31
+CGH_B 32
+CGH_E 33
+CGH_I 34
+CGH_S 35
+NSN 36
+NSN_B 37
+NSN_E 38
+NSN_I 39
+NSN_S 40
+SMK 41
+SMK_B 42
+SMK_E 43
+SMK_I 44
+SMK_S 45
+UHH 46
+UHH_B 47
+UHH_E 48
+UHH_I 49
+UHH_S 50
+UM 51
+UM_B 52
+UM_E 53
+UM_I 54
+UM_S 55
+3`_B 56
+3`_E 57
+3`_I 58
+3`_S 59
+4_B 60
+4_E 61
+4_I 62
+4_S 63
+@_B 64
+@_E 65
+@_I 66
+@_S 67
+@`_B 68
+@`_E 69
+@`_I 70
+@`_S 71
+A_B 72
+A_E 73
+A_I 74
+A_S 75
+D_B 76
+D_E 77
+D_I 78
+D_S 79
+E_B 80
+E_E 81
+E_I 82
+E_S 83
+E:_B 84
+E:_E 85
+E:_I 86
+E:_S 87
+I_B 88
+I_E 89
+I_I 90
+I_S 91
+N_B 92
+N_E 93
+N_I 94
+N_S 95
+O_B 96
+O_E 97
+O_I 98
+O_S 99
+OI_B 100
+OI_E 101
+OI_I 102
+OI_S 103
+S_B 104
+S_E 105
+S_I 106
+S_S 107
+T_B 108
+T_E 109
+T_I 110
+T_S 111
+U_B 112
+U_E 113
+U_I 114
+U_S 115
+V_B 116
+V_E 117
+V_I 118
+V_S 119
+Z_B 120
+Z_E 121
+Z_I 122
+Z_S 123
+aI_B 124
+aI_E 125
+aI_I 126
+aI_S 127
+aU_B 128
+aU_E 129
+aU_I 130
+aU_S 131
+b_B 132
+b_E 133
+b_I 134
+b_S 135
+d_B 136
+d_E 137
+d_I 138
+d_S 139
+dZ_B 140
+dZ_E 141
+dZ_I 142
+dZ_S 143
+eI_B 144
+eI_E 145
+eI_I 146
+eI_S 147
+f_B 148
+f_E 149
+f_I 150
+f_S 151
+g_B 152
+g_E 153
+g_I 154
+g_S 155
+h_B 156
+h_E 157
+h_I 158
+h_S 159
+i_B 160
+i_E 161
+i_I 162
+i_S 163
+j_B 164
+j_E 165
+j_I 166
+j_S 167
+k_B 168
+k_E 169
+k_I 170
+k_S 171
+l_B 172
+l_E 173
+l_I 174
+l_S 175
+m_B 176
+m_E 177
+m_I 178
+m_S 179
+n_B 180
+n_E 181
+n_I 182
+n_S 183
+oU_B 184
+oU_E 185
+oU_I 186
+oU_S 187
+p_B 188
+p_E 189
+p_I 190
+p_S 191
+r_B 192
+r_E 193
+r_I 194
+r_S 195
+s_B 196
+s_E 197
+s_I 198
+s_S 199
+t_B 200
+t_E 201
+t_I 202
+t_S 203
+tS_B 204
+tS_E 205
+tS_I 206
+tS_S 207
+u_B 208
+u_E 209
+u_I 210
+u_S 211
+v_B 212
+v_E 213
+v_I 214
+v_S 215
+w_B 216
+w_E 217
+w_I 218
+w_S 219
+z_B 220
+z_E 221
+z_I 222
+z_S 223
+{_B 224
+{_E 225
+{_I 226
+{_S 227
+#0 228
+#1 229
+#2 230
+#3 231
+#4 232
+#5 233
+#6 234
+#7 235
+#8 236
+#9 237
+#10 238
+#11 239
+#12 240
+#13 241
+#14 242
+#15 243
+#16 244
+#17 245
+#18 246
+#19 247
+#20 248
+#21 249
+#22 250
+#23 251
+#24 252
+#25 253
+#26 254
+#27 255
+#28 256
+#29 257
+#30 258
+#31 259
+#32 260
+#33 261
+#34 262
+#35 263
+#36 264
+#37 265
+#38 266
