+++ backend/model/ivector/global_cmvn.stats
@@ -1,0 +1,3 @@
+ [
+  2.74606e+11 7.912398e+09 -1.038969e+11 1.317063e+10 -5.916838e+10 -9.741645e+10 -2.601366e+10 -7.401961e+10 -2.684753e+10 -1.58957e+10 -4.96366e+10 -8.033801e+09 -3.074859e+10 -1.707706e+10 -1.639996e+10 -1.829851e+10 -1.09582e+10 -5.920052e+09 -7.745878e+09 -1.833566e+09 -2.802531e+09 -7.583578e+08 -3.994372e+08 6.940003e+07 8.935535e+08 2.282279e+08 1.745994e+09 -6.372821e+07 8.785109e+08 -1.030507e+08 8.091005e+08 4.407504e+08 1.859903e+09 1.753656e+09 2.066627e+09 1.37222e+09 1.386451e+09 -1.142544e+09 -2.51193e+08 -9.788437e+08 3.129497e+09 
+  2.54123e+13 1.928342e+12 6.751519e+12 1.637428e+12 3.016774e+12 5.912373e+12 1.868079e+12 3.886044e+12 1.841588e+12 1.633684e+12 2.137614e+12 1.072133e+12 1.296519e+12 1.00067e+12 8.062336e+11 6.591713e+11 4.603988e+11 3.121481e+11 2.243611e+11 1.153531e+11 6.293688e+10 2.146548e+10 2.735945e+09 7.957589e+08 1.053802e+10 2.59947e+10 4.301262e+10 5.558657e+10 7.239489e+10 8.546533e+10 8.436563e+10 7.792829e+10 8.569881e+10 9.576342e+10 7.780419e+10 5.335785e+10 5.166593e+10 4.813119e+10 3.81959e+10 2.395734e+10 0 ]
