+++ backend/model/ivector/splice.conf
@@ -1,0 +1,2 @@
+--left-context=3
+--right-context=3
